+++ routes/json/cover.json
@@ -1,14 +1,14 @@
 [
     {
-        "img": "https://miro.medium.com/max/442/1*-5x0MC7k53mV4VMqKiKHOg.png"
-    },
-    {
         "img": "https://miro.medium.com/max/6912/0*WAX1PSY7z2L3tOiZ"
     },
     {
-        "img": "https://miro.medium.com/max/2560/1*5yVOKthzCe2Ko5zoCgrLkQ.jpeg"
+        "img": "https://miro.medium.com/max/1460/1*evXUXDtSpBZo_HHfXt4CNA.jpeg"
     },
     {
-        "img": "https://miro.medium.com/max/1460/1*evXUXDtSpBZo_HHfXt4CNA.jpeg"
+        "img": "https://miro.medium.com/max/442/1*-5x0MC7k53mV4VMqKiKHOg.png"
+    },
+    {
+        "img": "https://miro.medium.com/max/2560/1*5yVOKthzCe2Ko5zoCgrLkQ.jpeg"
     }
 ]
