+++ views/blogs.ejs
@@ -17,9 +17,10 @@
                     <div class="news-item-info">
                         <div class="news-item-date button"><%= post.release %></div>
                         <h4 class="news-item-title item-title">
-                            <div class="container-fluid" style="background: white">
+                            <div class="container-fluid" style="background: white;">
                                 <a href="<%= post.href %>"
                                     title="<%= post.title %>"><%= post.title %></a>
+                            </div>
                         </h4>
                         <div class="news-item-desc">
                             <p>By <%= post.author %></p>
